Create buildguide for github actions
GITHUB_HEAD_REF: Only set for pull request events. The name of the head branch
GITHUB_REF: The branch or tag ref that triggered the workflow. For example, refs/heads/feature-branch-1. If neither a branch or tag is available for the event type, the variable will not exist.
GITHUB_RUN_ID	A unique number for each run within a repository. This number does not change if you re-run the workflow run.